experiments/s3-image-resize/requirements.txt
attrs==19.3.0
boto3==1.12.1
botocore==1.15.1
certifi==2019.11.28
chalice==1.12.0
chardet==3.0.4
Click==7.0
docutils==0.15.2
enum-compat==0.0.3
idna==2.8
jmespath==0.9.4
Pillow==7.0.0
python-dateutil==2.8.1
requests==2.22.0
requests-toolbelt==0.9.1
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.8
